+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for croud
 Unreleased
 ==========
 
+- Added a new ``print_raw`` function to simply print the output of croud.
+
 - Added new subcommand ``regions generate-deployment-manifest`` to fetch a deployment
   manifest for an edge region.